chore: Move PlayerManagerMixin to network directory
